Two additional tests to ensure invalid nonces are rejected.
The first test rejects an arbitraty nonce from the client, in this case the nonce is completely invalid - this could happen if a server is restarted and a new nonce manager created.
The second test verifies that a valid but previously used nonce is rejected - this covers the most likely replay attack scenario.