Update core/src/main/java/io/undertow/ajp/AjpResponseChannel.java
The header on the response end appears to be incorrect. With the header as is, Wireshark is unable to decode the exchange.

I do not expect that this would cause issues in most real-world scenarios at the moment, as the reuse flag is set to false. However, in reuse scenarios, I believe that some httpd daemons (including Apache HTTPD) may reject the end of the exchange, and close the connection.