Fixing substring-index in ProxyPeerAddress Handler
In ProxyPeerAddressHander, the substring boundary is wrongly calculated with 0 to index -1.
As the end-index is exclusive, it must be 0 to index.

Example XFF-Header without additional proxies (was correct before fix)
X-Forwarded-For: 123.123.123.123
Handler-Value: 123.123.123.123

Example XFF with at least one additional proxy:
X-Forwarded-For: 123.123.123.123, 210.210.210.210
Handler-Value (wrong): 123.123.123.12  (note the missing last character)
Handler-Value (fixed): 123.123.123.123