Parse languages according to RFC7231 / RFC5646 / BCP47
According to the newer standard RFC7231 (HTTP/1.1), the Accept-Language can
be used to indicate the desired languages by using language tags as defined
in RFC5646 (BCP 47).

This obsoletes the older standards RFC2616 and RFC2068, which had a different
definition for the content of Accept-Language.

Nevertheless, the newer definition is a superset of the old one, and therefore
the language tag parser should be changed to accept the newer BCP47 compatible
tags (which should not make any difference for almost all real world language
tags).