URLUtils.decode now will not decode spaces if space decoding is disabled
URLUtils.decode now will node decode all the spaces in URL even if this option was disabled using argument of the function. Decode function now will skip '%2(f/F)' sequences in decoding loop if space deconding is turned off.

The problem was that old version(1.3.5) worked fine, but after moving the new one the correct behaviour was broken. The bug was introduced for the edge case when our URL contains spaces '/' and their symbols '%2(f/F)' it tried to decode all the spaces event if this option was disabled and then encode them back (so it should look like spaces were not touched), but in fact in `mix` cases after decoding all spaces were encoded.
Example:
http://localhost/path%20with%20space/foo%2fbar --> http://localhost/path with spaces%2Ffoo%2Fbar while http://localhost/path with spaces/foo%2fbar is expected.

- bugfix
- tests