UNDERTOW-1267 add option to allow the modcluster advertise ttl to be set
Added variable including getter setter methods to allow the modcluster
advertise ttl to be set.  The default multicast ttl is now set to 10,
this matches the apache httpd mod_cluster behaviour.

Reordered several methods and variables so that their order matches.

Renamed variable group to address because it contains the advertise address,
not the advertise group.  Used the createUdpServer method as it requires
one less argument.