Fix duplicate extension initialization with multiple class loaders.
The Class object doesn't have equals or hashcode implemented. Therefore storing it in a
hashset won't help us to see if an extension has been initialized before or not.

This patch changes that check to be based on the Class name of the extension so the
handleDeployment method will be only called once per extension.