UNDERTOW-1640: Fix AbstractFramedStreamSinkChannel buffer access
When streams are canceled unexpectedly (http/2 rst_stream while blocking
operations execute, for example) interaction between close being called
on an I/O thread and blocking work on a task thread was not properly
handled.

Unlike the AbstractFramedStreamSourceChannel case I have not
forced a reproducer for invalid buffer access, it's a bit tricker
to hit without spending a lot of iterations, but the result would
be the same.