UNDERTOW-1664: Prevent errors when http/2 channels are terminated
Http/2 channels may be closed prior to the http2 server connection
object. In this case an error is no longer logged, and the exchange
is completed as though the connection is closed.

Exchange completion listeners are invoked in the unexpected case
to avoid resource leaks, and allow request logging to execute
as expected.