Http session to request/response references
Since 0c0abf2, http sessions keep reference to request and response via servletRequestContext.
Because an http session can be kept as a reference by the webapp for some time or for a long time after the end of the request, it would be better for memory that an http session does not keep references to request and response.