[UNDERTOW-1833] Unneccessary code removed from the test.
This is a trivial change of removal unused and unnecessary code in
the Http2EndExchangeTestCase.java test case.

Apart from that, I slightly changed the way how test fails in case that
HTTP2 connection is not correctly established - I unified this failure
with other failures in the test.

There is no other problem in the test as it is passing just fine.
I just got around this when I was investigating something else.