[UNDERTOW-1852] Add OSGi manifest headers to 3 Undertow modules
There's also "karaf" Maven module, enabled only with `osgi` Maven
profile. This is not to provide any additional artifact, but only to
verify the Manifest headers.

I also removed io.undertow.servlet.core.ServletExtensionHolder which was
supposed to provide a static way to add ServletExtensions in OSGi
environment. It's no longer needed and even in OSGi ServletExtensions
are now added normally to DeploymentInfo.