Initiate a graceful shutdown of each HTTP/2 connection
Use tracked HTTP/2 channels in GracefulShutdownHandler to
initate a graceful shutdown using GOAWAY according to RFC7540 6.8.
- Send initial GOAWAY with last stream id set to 2^31-1
- Send an updated GOAWAY with updated last stream id after 1 sec
- Force close if client has not already disconnected after 3 sec

Http/2 requests will be handled during the graceful shutdown,
but http/1 will be rejected with `503 Service Unavailable`

awaitShutdown() waits for both ongoing requests to finish and
for active HTTP/2 connections to be closed by the client.

Signed-off-by: Björn Svensson <bjorn.a.svensson@est.tech>