Add API to stop listening for new client connections
In addition to `suspend/resume of accepting new clients` its now also possible
to stop listening for new clients, i.e the listening socket will be closed.
This is needed when performing a graceful shutdown to avoid that clients just
reconnects via a new socket.

A code example of a graceful shutdown:

  // Stop listening for new client connections
  List<ListenerInfo> listeners = server.getListenerInfo();
  for (ListenerInfo listener : listeners) {
      listener.stop();
  }

  // Trigger a graceful shutdown of existing connections
  shutdownHandler.shutdown();

  // Wait for the graceful shutdown to finish
  shutdownHandler.awaitShutdown(6000);

  // Forceful shutdown of server
  server.stop();