Wire BlockHound into undertow tests
This defines a BlockHound integration within undertow-core, however
this could be an entirely separate module. For this to become
mergable, BlockHound would minimally have to become a `provided`
dependency so it's not added to consumers classpaths unnecessarily,
but applied in test modules

Arguably the blockhound integration belongs in an xnio-associated
module because xnio defines the relevant XnioIoThread class.

This idea is borrowed from Netty, which added BlockHound here:
https://github.com/netty/netty/pull/9687