[UNDERTOW-2075] Change a bit the execution of suspend resumes in AbstractFramedChannel: suspend is done asap (if in IO thread it's executed immediately, scheduled if not), resume is always postponed to the IO thread.
The boolean vars can be updated immediately but then we need synch all modifications, that update can be in IO thread or not, so synch is needed. We cannot do this in the freed method as we cannot synch on this. So for that case the var is also updated in the IO thread.

Signed-off-by: Flavia Rainone <frainone@redhat.com>