[UNDERTOW-2247] At HttpResponseConduit.processWrite, properly manage and clear the allocated buffer to prevent NPEs when ClosedChannelExceptions and broken pipe SocketExceptions occur.
Those NPEs can occur at the points where the buffer is closed and it is already null because the buffer was cleared by an inner recursive call as an attempt to deal with the exception.
The broken pipe can occur when a browser tab is closed during the response writing process.

Signed-off-by: Flavia Rainone <frainone@redhat.com>