[UNDERTOW-2255] Solve the Hpack decoding exception, by making sure that proxy handler does not forward http2 settings incompatible with the proxy local settings.
This was causing the proxy to send a bigger header table size to server in the HTTP2-Settings headers. As a consequence, the server would try to set the wrong table size, bigger than actually supported, and this would just result in an exception when hpack decoder is invoked with the new header table size.

Signed-off-by: Flavia Rainone <frainone@redhat.com>